+++ gen2-yolo/yolop/requirements.txt
@@ -4,3 +4,4 @@ torch>=1.7.0
 torchvision>=0.8.1
 numpy>=1.18.5
 blobconverter
+gdown
